microservices/audio_processing/python-service/requirements.txt
fastapi>=0.68.0,<1.0.0
uvicorn>=0.15.0,<1.0.0
python-multipart
openai-whisper
torch
torchaudio
soundfile
